python/requirements.txt
nats-py==2.8.0
protobuf==5.27.2
protobuf-serialization==0.1.2
pygob @ git+https://github.com/mgeisler/pygob@ebef7bcd2f13cad4975a9a06e43f6fbdfcb8b3fe
python-dateutil==2.9.0.post0
six==1.16.0
